requirements.txt
praw==7.7.1
nltk==3.8.1
pandas==2.1.4
matplotlib==3.8.2
wordcloud==1.9.3
python-dotenv==1.0.0 